+++ src/App.js
@@ -1,9 +1,9 @@
 import './App.css';
 import React, {useEffect} from "react";
 import {ThemeProvider} from "@mui/material";
-import RTL from "./theme/RTL";
+import RTL from "./Theme/RTL";
 import {Link, Outlet} from "react-router-dom";
-import {theme} from "./theme/rtl-theme";
+import {theme} from "./Theme/rtl-theme";
 import Api from "./Api";
